+++ packages/encryption/src/encryption/Decryption.ts
@@ -1,9 +1,7 @@
 import { unreachable } from '@dimensiondev/kit'
 import {
-    TypedMessage,
     decodeTypedMessageFromDocument,
     decodeTypedMessageV38ToV40Format,
-    ProfileIdentifier,
     AESCryptoKey,
     EC_Public_CryptoKey,
     andThenAsync,
@@ -11,6 +9,16 @@ import {
 import { None, Result } from 'ts-results'
 import { AESAlgorithmEnum, PayloadParseResult } from '../payload'
 import { decryptWithAES, importAESFromJWK } from '../utils'
+import {
+    DecryptionOption,
+    DecryptIO,
+    DecryptionProcess,
+    DecryptionProgressKind,
+    DecryptionError,
+    EphemeralECDH_Result,
+    DecryptionSuccess,
+} from './DecryptionTypes'
+const ErrorReasons = DecryptionError.Reasons
 
 export async function* decrypt(options: DecryptionOption, io: DecryptIO): AsyncIterableIterator<DecryptionProcess> {
     yield progress(DecryptionProgressKind.Started)
@@ -216,80 +224,5 @@ function progress(kind: DecryptionProgressKind.Success, rest: Omit<DecryptionSuc
 function progress(kind: DecryptionProgressKind.Started): DecryptionProcess
 function progress(kind: DecryptionProgressKind, rest?: object): DecryptionProcess {
     return { type: kind, ...rest } as any
-}
-
-export interface DecryptionOption {
-    message: PayloadParseResult.Payload
-    signal?: AbortSignal
-}
-export interface DecryptIO {
-    getPostKeyCache(): Promise<AESCryptoKey | null>
-    setPostKeyCache(key: AESCryptoKey): Promise<void>
-    hasLocalKeyOf(author: ProfileIdentifier): Promise<boolean>
-    /**
-     * Try to decrypt message by someone's localKey.
-     *
-     * Implementor must try authorHint's localKey if they have access to.
-     *
-     * Implementor may try other localKeys they owned even not listed in the authorHint.
-     *
-     * @param authorHint A hint for searching the localKey.
-     * @param data Encrypted data
-     * @param iv
-     * @returns The decrypted data
-     */
-    decryptByLocalKey(authorHint: ProfileIdentifier | null, data: Uint8Array, iv: Uint8Array): Promise<Uint8Array>
-    queryPublicKey(id: ProfileIdentifier | null, signal?: AbortSignal): Promise<EC_Public_CryptoKey>
-    queryPostKey_version40(iv: Uint8Array): Promise<StaticECDH_Result | null>
-    queryPostKey_version39(iv: Uint8Array, signal?: AbortSignal): AsyncIterableIterator<StaticECDH_Result>
-    queryPostKey_version38(iv: Uint8Array, signal?: AbortSignal): AsyncIterableIterator<StaticECDH_Result>
-    queryPostKey_version37(iv: Uint8Array, signal?: AbortSignal): AsyncIterableIterator<EphemeralECDH_Result>
-    /**
-     * Derive a group of AES key for ECDH.
-     *
-     * Implementor should derive a new AES-GCM key for each private key they have access to.
-     * @param publicKey The public key used in ECDH
-     * @returns This function MUST NOT throw an error. Error will be treated as fatal error.
-     *
-     * If the provided key cannot derive AES with any key (e.g. The given key is ED25519 but there is only P-256 private keys)
-     * please return an empty array.
-     */
-    deriveAESKey(publicKey: EC_Public_CryptoKey): Promise<AESCryptoKey[]>
-}
-export interface StaticECDH_Result {
-    encryptedKey: Uint8Array
-    iv: Uint8Array
-}
-export interface EphemeralECDH_Result extends StaticECDH_Result {
-    // It might be contained in the original payload.
-    ephemeralPublicKey?: EC_Public_CryptoKey
-    ephemeralPublicKeySignature?: Uint8Array
-}
-export enum DecryptionProgressKind {
-    Started = 'started',
-    Success = 'success',
-    Error = 'error',
-}
-export type DecryptionProcess = { type: DecryptionProgressKind.Started } | DecryptionSuccess | DecryptionError
-export interface DecryptionSuccess {
-    type: DecryptionProgressKind.Success
-    content: TypedMessage
-}
-enum ErrorReasons {
-    PayloadBroken = '[@masknet/encryption] Payload is broken.',
-    PayloadDecryptedButNoValidTypedMessageIsFound = '[@masknet/encryption] Payload decrypted, but no valid TypedMessage is found.',
-    CannotDecryptAsAuthor = '[@masknet/encryption] Failed decrypt as the author of this payload.',
-    DecryptFailed = '[@masknet/encryption] Post key found, but decryption failed.',
-    AuthorPublicKeyNotFound = "[@masknet/encryption] Cannot found author's public key",
-    MyPrivateKeyNotFound = '[@masknet/encryption] Cannot decrypt because there is no private key found.',
-    NotShareTarget = '[@masknet/encryption] No valid key is found. Likely this post is not shared with you',
-    Aborted = '[@masknet/encryption] Task aborted.',
-}
-export class DecryptionError extends Error {
-    static Reasons = ErrorReasons
-    readonly type = DecryptionProgressKind.Error
-    constructor(public override message: ErrorReasons, cause: unknown, public recoverable = false) {
-        super(message, { cause })
-    }
 }
 type Version = PayloadParseResult.Payload['version']
